content: generateTrackListDividedByGroups: inline util function
